Reorganiz files and names
